+++ MD/AtomParams.h
@@ -7,18 +7,18 @@ using namespace std;
 
 class AtomParams {
 	public:
-		AtomParams(float sig_, float eps_, float m_) : sig(sig_), eps(eps_), m(m_) {};
-		float sig;
-		float eps;
-		float m;
+		AtomParams(double sig_, double eps_, double m_) : sig(sig_), eps(eps_), m(m_) {};
+		double sig;
+		double eps;
+		double m;
 		int type;
 };
 //want a little class to make sure params get assigned types nicely
 class AtomParamWrapper {
 	public:
 		vector<AtomParams> params;
-		AtomParamWrapper(float rCut_) : rCut(rCut_) {}
-		const float rCut;
+		AtomParamWrapper(double rCut_) : rCut(rCut_) {}
+		const double rCut;
 		AtomParamWrapper &push_back(AtomParams params_) {
 			params_.type = params.size();
 			params.push_back(params_);
