+++ MD/Bounds.h
@@ -23,7 +23,7 @@ class Bounds : public BoundsGeneric {
 			return loStr + ", " + hiStr ;
 		}
 		bool atomInBounds(Atom *);
-		float volume();
+		double volume();
 	
 
 };
