+++ MD/FixLjCut.h
@@ -7,10 +7,10 @@
 using namespace std;
 class FixLjCut : public Fix {
 	public:
-		FixLjCut(vector<Atom *> &allAtoms_, vector<Atom *> atoms_, Data &simData_, InteractionParams &params_, float rCut) : Fix(allAtoms_, atoms_, simData_, 1), params(params_), rCutSqr(rCut*rCut) {};
+		FixLjCut(vector<Atom *> &allAtoms_, vector<Atom *> atoms_, Data &simData_, InteractionParams &params_, double rCut) : Fix(allAtoms_, atoms_, simData_, 1), params(params_), rCutSqr(rCut*rCut) {};
 		InteractionParams params;
 		void compute();
-		float rCutSqr;
+		double rCutSqr;
 };
